pyproject.toml
[project]
name = "newton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "aiosqlite>=0.21.0",
    "apscheduler>=4.0.0a5",
    "litellm>=1.81.12",
    "logfire>=4.24.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "pydantic-ai>=1.59.0",
    "pydantic-ai-litellm>=0.2.3",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.1",
    "sqlite-vec>=0.1.6",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
